requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
coverage==4.4.1
cryptography==2.0.3
Flask==0.12.2
Flask-JWT-Extended==3.3.1
Flask-Login==0.4.0
flask-mongoengine==0.9.3
Flask-Testing==0.6.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
mongoengine==0.13.0
pbr==3.1.1
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
pymongo==3.5.1
PyNaCl==1.1.2
PyYAML==3.12
six==1.10.0
Werkzeug==0.12.2
WTForms==2.1
